Avoid compute/store another fp on content if uuid is already the key. (#221)

+++ src/base/schema.rs
@@ -359,7 +359,7 @@ impl std::fmt::Display for FieldSchema {
 pub struct CollectorSchema {
     pub fields: Vec<FieldSchema>,
     /// If specified, the collector will have an automatically generated UUID field with the given index.
-    pub auto_uuid_field_idx: Option<usize>,
+    pub auto_uuid_field_idx: Option<u32>,
 }
 
 impl std::fmt::Display for CollectorSchema {
